feeback score can be bigger than SmallInteger limit, so up to BigInteger

+++ archive/models.py
@@ -33,7 +33,7 @@ class Item(models.Model):
                         max_length = 48 )
     iQuantity       = models.SmallIntegerField( 'quantity' )
     cSellerID       = models.CharField( 'seller user name', max_length = 48 )
-    iFeedbackScore  = models.SmallIntegerField( 'seller feedback score' )
+    iFeedbackScore  = models.BigIntegerField( 'seller feedback score' )
     cFeedbackPercent= models.CharField( 'seller feedback percent', max_length = 8 )
     iBidCount       = models.SmallIntegerField( 'bid count' )
     dConvertPrice   = models.DecimalField( # form was throwing nonsense error
